wechat-agent/pyproject.toml
[project]
name = "wechat-agent"
version = "0.1.0"
description = "wxauto message listener → NATS JetStream publisher"
requires-python = ">=3.13"
dependencies = [
    "nats-py>=2.10.0",
    "pydantic>=2.11.0",
    "python-json-logger>=3.3.0",
    "pyyaml>=6.0.2",
    "wxauto>=3.9.11.17",
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"

[project.scripts]
wechat-agent = "main:main"
